+++ Modules/Intent.Modules.Entities.Repositories.Api/Templates/EntityRepositoryInterface/EntityRepositoryInterfaceTemplatePartial.cs
@@ -5,6 +5,7 @@ using Intent.Modelers.Domain.Api;
 using Intent.Modules.Common;
 using Intent.Modules.Common.CSharp.Builder;
 using Intent.Modules.Common.CSharp.Templates;
+using Intent.Modules.Common.CSharp.VisualStudio;
 using Intent.Modules.Common.Templates;
 using Intent.Modules.Constants;
 using Intent.Modules.Entities.Repositories.Api.Templates.RepositoryInterface;
@@ -48,7 +49,7 @@ namespace Intent.Modules.Entities.Repositories.Api.Templates.EntityRepositoryInt
                             if (rootEntity.TryGetMetadata<CSharpProperty[]>("primary-keys", out var pks)
                                 && pks.Length == 1)
                             {
-                                @interface.AddMethod($"Task<{GetTypeName(TemplateFulfillingRoles.Domain.Entity.Interface, Model)}>", "FindByIdAsync", method =>
+                                @interface.AddMethod($"Task<{GetTypeName(TemplateFulfillingRoles.Domain.Entity.Interface, Model)}{(OutputTarget.GetProject().NullableEnabled ? "?" : "")}>", "FindByIdAsync", method =>
                                 {
                                     method.AddAttribute("[IntentManaged(Mode.Fully)]");
                                     var pk = pks.First();